test: don't reboot container during tests, move tests to unit
