+++ tests/integration/Core/Framework/ApiRoutesHaveASchemaTest.php
@@ -28,25 +28,17 @@ class ApiRoutesHaveASchemaTest extends TestCase
 
     private RouteCollection $routes;
 
-    private ContainerInterface $container;
-
     protected function setUp(): void
     {
-        $container = KernelLifecycleManager::bootKernel()->getContainer()->get('test.service_container');
-
-        static::assertInstanceOf(ContainerInterface::class, $container);
-
-        $this->container = $container;
-
-        $router = $this->container->get(RouterInterface::class);
+        $router = $this->getContainer()->get(RouterInterface::class);
         $this->routes = $router->getRouteCollection();
     }
 
     public function testStoreApiRoutesHaveASchema(): void
     {
-        $generator = $this->container->get(StoreApiGenerator::class);
+        $generator = $this->getContainer()->get(StoreApiGenerator::class);
         $schema = $generator->generate(
-            $this->container->get(SalesChannelDefinitionInstanceRegistry::class)->getDefinitions(),
+            $this->getContainer()->get(SalesChannelDefinitionInstanceRegistry::class)->getDefinitions(),
             DefinitionService::STORE_API,
             DefinitionService::TYPE_JSON_API,
             null
@@ -112,9 +104,9 @@ class ApiRoutesHaveASchemaTest extends TestCase
 
     public function testAdminApiRoutesHaveASchema(): void
     {
-        $generator = $this->container->get(OpenApi3Generator::class);
+        $generator = $this->getContainer()->get(OpenApi3Generator::class);
         $schema = $generator->generate(
-            $this->container->get(DefinitionInstanceRegistry::class)->getDefinitions(),
+            $this->getContainer()->get(DefinitionInstanceRegistry::class)->getDefinitions(),
             DefinitionService::API
         );
