+++ tests/unit/Core/Framework/App/Event/AppActivatedEventTest.php
@@ -1,9 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\Framework\App\Event\AppDeletedEvent;
+use Shopware\Core\Framework\App\AppEntity;
+use Shopware\Core\Framework\App\Event\AppActivatedEvent;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
@@ -11,29 +13,32 @@ use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 /**
  * @internal
  */
-class AppDeletedEventTest extends TestCase
+#[CoversClass(AppActivatedEvent::class)]
+class AppActivatedEventTest extends TestCase
 {
     public function testGetter(): void
     {
-        $appId = Uuid::randomHex();
+        $app = new AppEntity();
         $context = Context::createDefaultContext();
-        $event = new AppDeletedEvent(
-            $appId,
+        $event = new AppActivatedEvent(
+            $app,
             $context
         );
 
-        static::assertEquals($appId, $event->getAppId());
+        static::assertEquals($app, $event->getApp());
         static::assertEquals($context, $event->getContext());
-        static::assertSame(AppDeletedEvent::NAME, $event->getName());
-        static::assertSame(['keepUserData' => false], $event->getWebhookPayload());
+        static::assertSame(AppActivatedEvent::NAME, $event->getName());
+        static::assertSame([], $event->getWebhookPayload());
     }
 
     public function testIsAllowed(): void
     {
         $appId = Uuid::randomHex();
+        $app = (new AppEntity())
+            ->assign(['id' => $appId]);
         $context = Context::createDefaultContext();
-        $event = new AppDeletedEvent(
-            $appId,
+        $event = new AppActivatedEvent(
+            $app,
             $context
         );
