+++ tests/unit/Core/Framework/App/Event/AppDeactivatedEventTest.php
@@ -1,12 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\AppEntity;
-use Shopware\Core\Framework\App\Event\AppInstalledEvent;
-use Shopware\Core\Framework\App\Event\AppUpdatedEvent;
-use Shopware\Core\Framework\App\Manifest\Manifest;
+use Shopware\Core\Framework\App\Event\AppDeactivatedEvent;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
@@ -14,24 +13,22 @@ use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 /**
  * @internal
  */
-class AppUpdatedEventTest extends TestCase
+#[CoversClass(AppDeactivatedEvent::class)]
+class AppDeactivatedEventTest extends TestCase
 {
     public function testGetter(): void
     {
         $app = new AppEntity();
         $context = Context::createDefaultContext();
-        $event = new AppUpdatedEvent(
+        $event = new AppDeactivatedEvent(
             $app,
-            Manifest::createFromXmlFile(__DIR__ . '/../Manifest/_fixtures/test/manifest.xml'),
             $context
         );
 
         static::assertEquals($app, $event->getApp());
         static::assertEquals($context, $event->getContext());
-        static::assertSame(AppUpdatedEvent::NAME, $event->getName());
-        static::assertSame([
-            'appVersion' => '1.0.0',
-        ], $event->getWebhookPayload());
+        static::assertSame(AppDeactivatedEvent::NAME, $event->getName());
+        static::assertSame([], $event->getWebhookPayload());
     }
 
     public function testIsAllowed(): void
@@ -40,9 +37,8 @@ class AppUpdatedEventTest extends TestCase
         $app = (new AppEntity())
             ->assign(['id' => $appId]);
         $context = Context::createDefaultContext();
-        $event = new AppInstalledEvent(
+        $event = new AppDeactivatedEvent(
             $app,
-            Manifest::createFromXmlFile(__DIR__ . '/../Manifest/_fixtures/test/manifest.xml'),
             $context
         );
