+++ tests/unit/Core/Framework/App/Event/AppDeletedEventTest.php
@@ -1,10 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\Framework\App\AppEntity;
-use Shopware\Core\Framework\App\Event\AppActivatedEvent;
+use Shopware\Core\Framework\App\Event\AppDeletedEvent;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
@@ -12,31 +12,30 @@ use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 /**
  * @internal
  */
-class AppActivatedEventTest extends TestCase
+#[CoversClass(AppDeletedEvent::class)]
+class AppDeletedEventTest extends TestCase
 {
     public function testGetter(): void
     {
-        $app = new AppEntity();
+        $appId = Uuid::randomHex();
         $context = Context::createDefaultContext();
-        $event = new AppActivatedEvent(
-            $app,
+        $event = new AppDeletedEvent(
+            $appId,
             $context
         );
 
-        static::assertEquals($app, $event->getApp());
+        static::assertEquals($appId, $event->getAppId());
         static::assertEquals($context, $event->getContext());
-        static::assertSame(AppActivatedEvent::NAME, $event->getName());
-        static::assertSame([], $event->getWebhookPayload());
+        static::assertSame(AppDeletedEvent::NAME, $event->getName());
+        static::assertSame(['keepUserData' => false], $event->getWebhookPayload());
     }
 
     public function testIsAllowed(): void
     {
         $appId = Uuid::randomHex();
-        $app = (new AppEntity())
-            ->assign(['id' => $appId]);
         $context = Context::createDefaultContext();
-        $event = new AppActivatedEvent(
-            $app,
+        $event = new AppDeletedEvent(
+            $appId,
             $context
         );
