+++ tests/unit/Core/Framework/App/Event/AppFlowActionEventTest.php
@@ -1,7 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\Event\AppFlowActionEvent;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
@@ -9,6 +10,7 @@ use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 /**
  * @internal
  */
+#[CoversClass(AppFlowActionEvent::class)]
 class AppFlowActionEventTest extends TestCase
 {
     public function testGetter(): void
