+++ tests/unit/Core/Framework/App/Event/AppUpdatedEventTest.php
@@ -1,10 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Tests\Integration\Core\Framework\App\Event;
+namespace Shopware\Tests\Unit\Core\Framework\App\Event;
 
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\AppEntity;
-use Shopware\Core\Framework\App\Event\AppDeactivatedEvent;
+use Shopware\Core\Framework\App\Event\AppUpdatedEvent;
+use Shopware\Core\Framework\App\Manifest\Manifest;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
@@ -12,21 +14,25 @@ use Shopware\Core\Framework\Webhook\AclPrivilegeCollection;
 /**
  * @internal
  */
-class AppDeactivatedEventTest extends TestCase
+#[CoversClass(AppUpdatedEvent::class)]
+class AppUpdatedEventTest extends TestCase
 {
     public function testGetter(): void
     {
         $app = new AppEntity();
         $context = Context::createDefaultContext();
-        $event = new AppDeactivatedEvent(
+        $event = new AppUpdatedEvent(
             $app,
+            Manifest::createFromXmlFile(__DIR__ . '/../Manifest/_fixtures/test/manifest.xml'),
             $context
         );
 
         static::assertEquals($app, $event->getApp());
         static::assertEquals($context, $event->getContext());
-        static::assertSame(AppDeactivatedEvent::NAME, $event->getName());
-        static::assertSame([], $event->getWebhookPayload());
+        static::assertSame(AppUpdatedEvent::NAME, $event->getName());
+        static::assertSame([
+            'appVersion' => '1.0.0',
+        ], $event->getWebhookPayload());
     }
 
     public function testIsAllowed(): void
@@ -35,8 +41,9 @@ class AppDeactivatedEventTest extends TestCase
         $app = (new AppEntity())
             ->assign(['id' => $appId]);
         $context = Context::createDefaultContext();
-        $event = new AppDeactivatedEvent(
+        $event = new AppUpdatedEvent(
             $app,
+            Manifest::createFromXmlFile(__DIR__ . '/../Manifest/_fixtures/test/manifest.xml'),
             $context
         );
